Restructured yet again my actions and reducers folder and changed it to the original state. Still have bugs that needs to be addressed-WIP
